修改GPU资源类的基类

+++ Engine/Source/RHI/Private/OpenGL/OpenGLVertexBufferResource.h
@@ -26,7 +26,7 @@ namespace Panda
 
         void Bind() override;
 
-        int GetVertexCount() const { return DataSize / GetLayout().GetStride(); }
+        size_t GetVertexCount() const { return DataSize / GetLayout().GetStride(); }
 
     private:
         int CurrentOffset = 0;
